+++ src/Raven.CodeAnalysis/Macros/IMacroKeywordProvider.cs
@@ -1,0 +1,8 @@
+using System.Collections.Immutable;
+
+namespace Raven.CodeAnalysis.Macros;
+
+public interface IMacroKeywordProvider
+{
+    ImmutableArray<MacroKeyword> Keywords { get; }
+}