+++ src/Raven.CodeAnalysis/Macros/IMacroTokenStreamProvider.cs
@@ -1,0 +1,6 @@
+namespace Raven.CodeAnalysis.Macros;
+
+public interface IMacroTokenStreamProvider
+{
+    IMacroTokenStream CreateTokenStream(MacroTokenStreamContext context);
+}